# specs/001-backend-system/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Internal Bank Reconciliation System API
  description: |
    Enterprise-grade API for bank transaction reconciliation.
    Supports CSV upload, automated matching, manual overrides, and approval workflows.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@ibrs.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.ibrs.example.com/v1
    description: Production server
  - url: https://staging-api.ibrs.example.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    ReconciliationSession:
      type: object
      required:
        - id
        - fileName
        - fileSize
        - uploadedAt
        - uploadedBy
        - totalBankTransactions
        - totalInternalTransactions
        - matchedCount
        - unmatchedBankOnlyCount
        - unmatchedInternalOnlyCount
        - varianceDetectedCount
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        uploadedAt:
          type: string
          format: date-time
        uploadedBy:
          type: string
          format: uuid
        totalBankTransactions:
          type: integer
        totalInternalTransactions:
          type: integer
        matchedCount:
          type: integer
        unmatchedBankOnlyCount:
          type: integer
        unmatchedInternalOnlyCount:
          type: integer
        varianceDetectedCount:
          type: integer
        status:
          $ref: '#/components/schemas/ReconciliationStatus'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        approvedBy:
          type: string
          format: uuid
          nullable: true
        rejectedAt:
          type: string
          format: date-time
          nullable: true
        rejectedBy:
          type: string
          format: uuid
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        processingStartedAt:
          type: string
          format: date-time
          nullable: true
        processingCompletedAt:
          type: string
          format: date-time
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReconciliationStatus:
      type: string
      enum:
        - draft
        - processing
        - pending_approval
        - approved
        - rejected
...
    ReconciliationSummary:
      type: object
      required:
        - totalTransactions
        - matchedPercentage
        - unmatchedPercentage
        - variancePercentage
        - totalVarianceAmount
      properties:
        totalTransactions:
          type: integer
        matchedPercentage:
          type: number
          format: float
        unmatchedPercentage:
          type: number
          format: float
        variancePercentage:
          type: number
          format: float
        totalVarianceAmount:
          type: number
          format: double
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalRows
        - totalPages
        - hasNextPage
        - hasPreviousPage
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalRows:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean

    # === Error Response ===
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              nullable: true
              description: Field-specific validation errors
            stack:
              type: string
              nullable: true
              description: Stack trace (development mode only)
  parameters:
...
    reconciliationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Reconciliation session ID
...
    pageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number
    pageSizeParam:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
      description: Items per page
...
paths:
...
  /reconciliations:
    get:
      tags:
        - Reconciliations
      summary: List reconciliations
      description: Returns paginated list of reconciliation sessions
      operationId: listReconciliations
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - name: status
          in: query
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ReconciliationStatus'
          style: form
          explode: true
          description: Filter by status
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter by upload date (from)
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter by upload date (to)
      responses:
        '200':
          description: List of reconciliations
          content:
            application/json:
              schema:
                type: object
                required:
                  - reconciliations
                  - pagination
                properties:
                  reconciliations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReconciliationSession'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Reconciliations
      summary: Create reconciliation
      description: Upload bank CSV file and initiate reconciliation
      operationId: createReconciliation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file containing bank transactions
      responses:
        '200':
          description: Reconciliation created
          content:
            application/json:
              schema:
                type: object
                required:
                  - session
                properties:
                  session:
                    $ref: '#/components/schemas/ReconciliationSession'
        '400':
          description: Invalid file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reconciliations/{id}:
    get:
      tags:
        - Reconciliations
      summary: Get reconciliation
      description: Returns reconciliation session details with summary
      operationId: getReconciliation
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/reconciliationId'
      responses:
        '200':
          description: Reconciliation details
          content:
            application/json:
              schema:
                type: object
                required:
                  - session
                  - summary
                properties:
                  session:
                    $ref: '#/components/schemas/ReconciliationSession'
                  summary:
                    $ref: '#/components/schemas/ReconciliationSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reconciliation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []